@@ -34,28 +34,28 @@
   },
   "homepage": "https://github.com/Horat1us/react-hooks#readme",
   "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.10.4",
-    "@rollup/plugin-babel": "^5.0.4",
-    "@rollup/plugin-node-resolve": "^8.1.0",
+    "@babel/core": "^7.26.0",
+    "@babel/plugin-transform-object-rest-spread": "^7.25.9",
+    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-node-resolve": "^15.3.0",
     "@types/jest": "^26.0.3",
-    "@types/react": "^16.9.41",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-test-renderer": "^16.9.2",
-    "jest": "^26.1.0",
-    "jsdom": "^16.2.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "rollup": "^1.32.1",
-    "ts-jest": "^26.1.1",
-    "typescript": "^3.9.6"
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jsdom": "^25.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "rollup": "^4.28.1",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.7.2"
   },
   "peerDependencies": {
-    "react": "^16.8.0 || ^17.0.0 || ^18.0.0"
+    "react": "^16.8.0 || ^17.0.0 || ^18.0.0 || ^19.0.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.10.4"
