@@ -1,15 +1,16 @@
 import * as React from "react";
-import * as ReactDOM from "react-dom";
-import {act} from "react-dom/test-utils";
+import { createRoot } from 'react-dom/client';
+import {act} from 'react';
 import {DocumentTitle} from "../../src/components";
+
+(global as any).IS_REACT_ACT_ENVIRONMENT = true
 
 test("<DocumentTitle />", () => {
     const newTitle = `<DocumentTitle> test case`;
     act(() => {
-        ReactDOM.render(<DocumentTitle data-x="s">{newTitle}</DocumentTitle>, document.head);
+        const root = createRoot(document.head);
+        root.render(<DocumentTitle data-x="s">{newTitle}</DocumentTitle>);
     });
     expect(document.title).toEqual(newTitle);
     expect(document.head).toMatchSnapshot();
 });
-
-
