@@ -1,6 +1,6 @@
 import babel from "@rollup/plugin-babel";
 import resolve from "@rollup/plugin-node-resolve";
-import pkg from './package.json';
+import pkg from './package.json' with {type: "json"};
 
 const extensions = [".ts", ".tsx",];
 const external = [
