@@ -1,8 +1,8 @@
 import * as React from "react";
 
-export const openStateReducer: React.Reducer<boolean, any> = (prevState) => !prevState;
+export const openStateReducer = (prevState: boolean): boolean => !prevState;
 export const useOpenState = (initialState: boolean = false): [boolean, () => void] => {
-    return React.useReducer<React.Reducer<boolean, any>>(
+    return React.useReducer<boolean, []>(
         openStateReducer,
         initialState
     ) as [boolean, () => void];
